fix(types): type Pine ints as int statically and correct measured result types
A Pine int is a double at runtime, but its static type was float, so a type
checker rejected every int-typed value handed to an int parameter and every
int literal used where Pine expects one. PyneInt and na_int are now int in the
stubs, while the runtime representation stays a float.

- types: new pine_int bridge. At runtime it is float itself; the stub types its
  result as PyneInt. Lib functions that return a Pine int (time parts,
  timestamp, sizes and indices, strategy trade counters, cast_int, safe_int,
  input.int) return through it instead of a bare float().
- core/pine_range.pyi: the for-loop counter is typed from its bounds (int
  bounds give an int counter, a float bound a float one).
- Generic results follow their source: nz, fixnan, request.seed, the generic
  input(), math.abs/max/min, ta.valuewhen/median/range, and the array and
  matrix reductions (an int array reduces to an int). Matrix functions keep the
  element type. inline_series returns the value type instead of SeriesImpl.
- Result types corrected after compile-only TradingView measurements:
  color.r/g/b, math.sum/avg/sign and the ta.supertrend direction are float;
  timenow, dividends.future_ex_date/future_pay_date and
  syminfo.shares_outstanding_total are int; syminfo.target_price_estimates is
  float; ta.valuewhen/median/range/mode take their source's type. The runtime
  type inference follows (LIB_TYPE_OVERRIDES, regenerated lib_types.json and
  module_properties.json), and two inference tests now expect the measured
  float for math.sign and the supertrend direction.

+++ src/pynecore/types/na.pyi
@@ -81,9 +81,10 @@ class NA(Generic[T]):
 # The singleton constants wear the same "plain T" face as __new__ above:
 # na_float must be assignable wherever a float is, or every hot-path use of the
 # interned constants (instead of an NA(float) call) would false-positive.
-# (na_float genuinely IS a float now — the interned native nan.)
+# (na_float genuinely IS a float now — the interned native nan.) na_int is the
+# same nan at runtime, but it is the na of a Pine int, so it types as one.
 na_float: float
-na_int: float
+na_int: int
 na_str: str
 na_bool: bool
